+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ### Changes
 
-- Kotlin 1.2.20
+- Kotlin 1.2.21
 - AnchorPaneConstraint properties now accept any Number, not just Double
 - AbstractField.textProperty was renamed to labelProperty to avoid confusion with the textProperty() exposed by textfields inside of a field
 - ItemViewModel.bind `defaultValue` parameter